Renaming varaibles and updating to latest video observer.
